requirements.txt
tensorflow-cpu==2.20.0
pillow==11.3.0
Flask==3.1.2
imageio==2.37.0
